Refactors all plugins to use actionableMessageEmitter, a custom event emitter. It comes in two flavors:
actionableMessage 		- any incoming message that is not from a hostmask that is ignored

actionableMessageAddressingBot  - any incoming message that is not from a hostmask that is ignored,
			 	  and addressing the bot by its nick.

The effect is that instead of running a database query to check if a hostmask is ignored for each plugin,
each plugin can instead listen to a custom event. This means that we're reducing the number of queries
per message to one instead of one query for each plugin (there are around 23 plugins at the time of this writing!)

These changes should improve performance significantly, due to eliminating the aforementioned overhead.